backend/foodgram/requirements.txt
Django == 4.1.7
django-colorfield==0.8.0
django-filter==23.1
djangorestframework==3.14.0
djoser==2.1.0
drf_extra_fields==3.5.0
flake8==6.0.0
flake8-isort==6.0.0
flake8-return==1.2.0
gunicorn==20.1.0
pep8-naming==0.13.3
# Pillow==9.4.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0